Move some mock utilities into mocks folder
