+++ tests/mocks/actions.js
@@ -1,0 +1,53 @@
+const { testRoot } = require("../utils");
+const { defaultActionInfo } = require("./github");
+
+const prBenchName = "local-framework";
+const baseBenchName = "base-framework";
+const defaultInputs = Object.freeze({
+	path: testRoot("results/test-results.json"),
+	reportId: null,
+	prBenchName,
+	baseBenchName,
+	defaultOpen: false,
+	keepOldResults: false,
+});
+
+/** @type {import('../../src/global').Logger} */
+const testLogger = {
+	debug() {},
+	info() {},
+	warn() {},
+	startGroup() {},
+	endGroup() {},
+};
+
+/**
+ * @type {import('../../src/global').GitHubActionContext}
+ */
+const fakeGitHubContext = {
+	payload: null,
+	eventName: "pull_request",
+	sha: "context.sha",
+	ref: "context.ref",
+	workflow: defaultActionInfo.workflow.name,
+	action: "context.action",
+	actor: "context.actor",
+	job: "context.job",
+	runNumber: defaultActionInfo.run.number,
+	runId: defaultActionInfo.run.id,
+	repo: {
+		owner: "andrewiggins",
+		repo: "tachometer-reporter-action",
+	},
+	issue: {
+		owner: "andrewiggins",
+		repo: "tachometer-reporter-action",
+		number: 5,
+	},
+};
+
+module.exports = {
+	defaultInputs,
+	fakeGitHubContext,
+	testLogger,
+};
